requirements.txt
fastapi
uvicorn
sqlalchemy
uuid
pydantic
psycopg2
psycopg2-binary
python-dotenv
slowapi
aioredis=1.3.1